Skapat vissa exceptions + vissa mapperklasser + uppdaterat på vissa ställen + gjort endpoint för Lista filmer på kund och börjat på Reservera lokal för kund

+++ src/main/java/com/leander/cinema/repository/TicketRepository.java
@@ -3,5 +3,5 @@ package com.leander.cinema.repository;
 import com.leander.cinema.entity.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 }
